Move TestConfig to tests package
